Update Capacitor calls
