+++ api/web/package.json
@@ -31,6 +31,7 @@
     "dependencies": {
         "@capacitor-firebase/messaging": "^8.2.0",
         "@capacitor/browser": "^8.0.1",
+        "@capacitor/camera": "^8.2.0",
         "@capacitor/clipboard": "^8.0.1",
         "@capacitor/core": "^8.3.1",
         "@capacitor/geolocation": "^8.1.0",
