+++ api/web/src/base/capacitor.ts
@@ -1,8 +1,5 @@
 import { Browser } from '@capacitor/browser';
-import { Clipboard } from '@capacitor/clipboard';
 import { Capacitor } from '@capacitor/core';
-import { Geolocation } from '@capacitor/geolocation';
-import type { CallbackID, Position, PositionOptions } from '@capacitor/geolocation';
 
 export function isNativePlatform(): boolean {
     return Capacitor.isNativePlatform();
@@ -10,10 +7,6 @@ export function isNativePlatform(): boolean {
 
 export function supportsServiceWorker(): boolean {
     return typeof navigator !== 'undefined' && !isNativePlatform() && 'serviceWorker' in navigator;
-}
-
-export function supportsLocationRequests(): boolean {
-    return isNativePlatform() || (typeof navigator !== 'undefined' && 'geolocation' in navigator);
 }
 
 export async function openExternalUrl(url: string | URL): Promise<void> {
@@ -43,56 +36,4 @@ export async function openSecondaryView(url: string | URL): Promise<void> {
     }
 
     window.open(href.toString(), '_blank', 'noopener');
-}
-
-export async function writeClipboardText(value: string): Promise<void> {
-    await Clipboard.write({ string: value });
-}
-
-function normalizeNativePermissionState(state: string | null | undefined): PermissionState | 'prompt' | 'unknown' {
-    switch (state) {
-        case 'granted':
-        case 'denied':
-            return state;
-        case 'prompt':
-        case 'prompt-with-rationale':
-            return 'prompt';
-        default:
-            return 'unknown';
-    }
-}
-
-export async function checkNativeLocationPermission(): Promise<PermissionState | 'prompt' | 'unknown'> {
-    try {
-        const status = await Geolocation.checkPermissions();
-        return normalizeNativePermissionState(status.location ?? status.coarseLocation);
-    } catch (err) {
-        console.warn('Failed to query native geolocation permission status', err);
-        return 'unknown';
-    }
-}
-
-export async function requestNativeLocationPermission(): Promise<PermissionState | 'prompt' | 'unknown'> {
-    try {
-        const status = await Geolocation.requestPermissions();
-        return normalizeNativePermissionState(status.location ?? status.coarseLocation);
-    } catch (err) {
-        console.warn('Failed to request native geolocation permission', err);
-        return 'unknown';
-    }
-}
-
-export async function getCurrentLocation(options?: PositionOptions): Promise<Position> {
-    return Geolocation.getCurrentPosition(options);
-}
-
-export async function watchLocation(
-    options: PositionOptions,
-    callback: (position: Position | null, err?: unknown) => void
-): Promise<CallbackID> {
-    return Geolocation.watchPosition(options, callback);
-}
-
-export async function clearLocationWatch(id: CallbackID): Promise<void> {
-    await Geolocation.clearWatch({ id });
 }
