+++ api/web/src/stores/device/geolocation.ts
@@ -1,12 +1,60 @@
-import {
-    checkNativeLocationPermission,
-    getCurrentLocation,
-    isNativePlatform,
-    requestNativeLocationPermission,
-    supportsLocationRequests
-} from '../../base/capacitor.ts';
+import { Geolocation } from '@capacitor/geolocation';
+import type { CallbackID, Position, PositionOptions } from '@capacitor/geolocation';
+import { isNativePlatform } from '../../base/capacitor.ts';
 import { queryPermissionStatus } from './shared.ts';
 import type { DevicePermissionContext } from './types.ts';
+
+function normalizeNativeLocationPermission(state: string | null | undefined): PermissionState | 'prompt' | 'unknown' {
+    switch (state) {
+        case 'granted':
+        case 'denied':
+            return state;
+        case 'prompt':
+        case 'prompt-with-rationale':
+            return 'prompt';
+        default:
+            return 'unknown';
+    }
+}
+
+export function supportsLocationRequests(): boolean {
+    return isNativePlatform() || (typeof navigator !== 'undefined' && 'geolocation' in navigator);
+}
+
+export async function checkNativeLocationPermission(): Promise<PermissionState | 'prompt' | 'unknown'> {
+    try {
+        const status = await Geolocation.checkPermissions();
+        return normalizeNativeLocationPermission(status.location ?? status.coarseLocation);
+    } catch (err) {
+        console.warn('Failed to query native geolocation permission status', err);
+        return 'unknown';
+    }
+}
+
+export async function requestNativeLocationPermission(): Promise<PermissionState | 'prompt' | 'unknown'> {
+    try {
+        const status = await Geolocation.requestPermissions();
+        return normalizeNativeLocationPermission(status.location ?? status.coarseLocation);
+    } catch (err) {
+        console.warn('Failed to request native geolocation permission', err);
+        return 'unknown';
+    }
+}
+
+export async function getCurrentLocation(options?: PositionOptions): Promise<Position> {
+    return Geolocation.getCurrentPosition(options);
+}
+
+export async function watchLocation(
+    options: PositionOptions,
+    callback: (position: Position | null, err?: unknown) => void
+): Promise<CallbackID> {
+    return Geolocation.watchPosition(options, callback);
+}
+
+export async function clearLocationWatch(id: CallbackID): Promise<void> {
+    await Geolocation.clearWatch({ id });
+}
 
 export class GeolocationPermission {
     constructor(private readonly context: DevicePermissionContext) {}
